requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy==2.0.25
pydantic==2.5.3
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
python-multipart==0.0.6
stripe==7.11.0
fal-client==0.4.1
replicate==0.23.1
python-dotenv==1.0.0
pydantic[email]
bcrypt
httpx 
librosa 
soundfile
pyotp
qrcode
Pillow
alembic
disposable-email-domains
cryptography
apscheduler
psycopg2-binary
python-dateutil
google-analytics-data